requirements.txt
# Core Python
python>=3.11

# Web Framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
streamlit==1.29.0

# Data Science
numpy==1.24.3
scipy==1.11.4
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
scipy  # For wavfile

# Async Support
asyncio==3.4.3
aiohttp==3.9.1
python-socketio==5.10.0

# Telegram Bot
python-telegram-bot==20.7

# Image Processing
pillow==10.1.0
opencv-python-headless==4.8.1.78

# HTTP/API
requests==2.31.0
httpx==0.25.2

# Database (optional)
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Environment
python-dotenv==1.0.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0

# Deployment
gunicorn==21.2.0
uvicorn==0.25.0
whitenoise==6.6.0

# Optional: Quantum
# qiskit==0.45.1
# qiskit-aer==0.13.1